requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.3
ed25519==1.5
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
magic-filter==1.0.12
Markdown==3.7
mistune==3.0.2
multidict==6.1.0
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
rsa==4.9
six==1.16.0
sniffio==1.3.1
Telethon==1.38.1
TgCrypto==1.2.5
Token==1.8.6
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.0
zipp==3.21.0
